Word is now translated when selecting the word. The translate UI still has to be made
I have experimented with the translation being shown in the 'copy/paste' bar, but the resulted in the menu needing over 10 seconds (in some occasions) to show up, while the translation was logged almost instantly. This is unacceptable, so the new approach is that the translation will show somewhere else as soon as a word is selected.